+++ wikipron/config.py
@@ -6,10 +6,10 @@ import re
 from typing import Callable, Optional
 
 import iso639
-import segments
 
-from wikipron.languagecodes import LANGUAGE_CODES
 from wikipron.extract import EXTRACTION_FUNCTIONS
+from wikipron.extract.default import extract_word_pron_default
+from wikipron.languagecodes import LANGUAGE_CODES
 from wikipron.typing import ExtractFunc, Pron, Word
 
 # GH-49: Estonian and Slovak use @title = "wikipedia:{language} phonology".
@@ -116,6 +116,10 @@ class Config:
     def _get_process_pron(
         self, no_stress: bool, no_syllable_boundaries: bool, no_segment: bool
     ) -> Callable[[Pron], Pron]:
+        # segments v2.1.2 oddly sets a global logging configuration
+        # that interferes with downstream logging.
+        # See: https://github.com/cldf/segments/issues/47
+        import segments
         processors = []
         if no_stress:
             processors.append(functools.partial(re.sub, r"[ˈˌ]", ""))
@@ -157,4 +161,4 @@ class Config:
         try:
             return EXTRACTION_FUNCTIONS[language]
         except KeyError:
-            return EXTRACTION_FUNCTIONS["default"]
+            return extract_word_pron_default