+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Unreleased
 
 ### Added
 
+-   Handled Latin, for which the actual graphemes cannot be the Wiktionary
+    page titles and have to come from within the page. (\# 92)
 -   Handled Thai, whose pronunciations are embedded in HTML tables. (\#90)
 -   Handled Khmer, whose pronunciations are embedded in HTML tables. (\#88)
 -   IPA segmentation using spaces by default, with the `--no-segment` flag to